{/src/main/java/com/panelist_rm/panelist/cns_panelist_datamodel.java}

+++ src/main/java/com/panelist_rm/panelist/cns_panelist_datamodel.java
@@ -9,7 +9,7 @@ public class cns_panelist_datamodel implements java.io.Serializable {
 
 	static final long serialVersionUID = 1L;
 
-	private java.util.List<java.lang.Integer> multi_choice_codes_05;
+	private java.util.List<java.lang.Long> multi_choice_codes_05;
 
 	public cns_panelist_datamodel() {
 	}
@@ -3155,12 +3155,12 @@ public class cns_panelist_datamodel implements java.io.Serializable {
 		this.end_of_rec = end_of_rec;
 	}
 
-	public java.util.List<java.lang.Integer> getMulti_choice_codes_05() {
+	public java.util.List<java.lang.Long> getMulti_choice_codes_05() {
 		return this.multi_choice_codes_05;
 	}
 
 	public void setMulti_choice_codes_05(
-			java.util.List<java.lang.Integer> multi_choice_codes_05) {
+			java.util.List<java.lang.Long> multi_choice_codes_05) {
 		this.multi_choice_codes_05 = multi_choice_codes_05;
 	}
